backend/requirements.txt
fastapi>=0.119.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
fastapi-mail==1.4.1
email-validator==2.1.0
reportlab==4.0.8
cloudinary==1.36.0
groq
pydantic>=2.7.0,<3.0
bcrypt==4.0.1
passlib>=1.7.4
openpyxl==3.1.2
